backend/requirements.txt
Flask
Flask-PyMongo
Flask-JWT-Extended
Flask-CORS
Werkzeug
gridfs
bson
PyMongo